we did some stuff!
"wireframing" (if you can even call it that) of the handle user page. We need css!!!

Firebase - add a way to get all users from the database. This is probably a _terrible_ implementation, if it even works. Technically, this should be streamed and paginated (What if we have a million users???,) but I'm not sure how we do that.

Firebase - We need a way to create & add a user to the database, please code for it!